+++ app/pages/settings.vue
@@ -8,10 +8,8 @@
                 {{ $t("settingsChangePassword") }}
             </h2>
             <PasswordForm />
-
             <ApiList />
         </ContentWrapper>
-        <AppFooter />
     </div>
 </template>
 <script setup lang="ts">